requirements.txt
torch==2.0.0
loguru==0.5.3
pyyaml==6.0
torchmetrics==0.11.4
tensorboardx==2.6.0
tensorflow-gpu==2.8.0